requirements.txt
black==22.6.0
boto3==1.25.0
botocore==1.28.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudscraper==1.2.60
garminconnect==0.1.45
idna==3.3
jmespath==1.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
requests-toolbelt==0.9.1
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.11
